fix: protocols on different sessisons
